+++ CHANGES.md
@@ -39,6 +39,8 @@ CUPS v2.4rc1 (Pending)
 - Now show better error messages when a driver interface program fails to
   provide a PPD file (Issue #148)
 - Added a workaround for Solaris in `httpAddrConnect2` (Issue #156)
+- Fixed an interaction between `--remote-admin` and `--remote-any` for the
+  `cupsctl` command (Issue #158)
 - Now use a 60 second timeout for reading USB backchannel data (Issue #160)
 - The USB backend now tries harder to find a serial number (Issue #170)
 - Fixed `@IF(name)` handling in `cupsd.conf` (Apple #5918)